Feature/foods page

+++ db/migrate/20230213135552_create_food.rb
@@ -3,7 +3,7 @@ class CreateFood < ActiveRecord::Migration[7.0]
     create_table :foods do |t|
       t.string :name
       t.string :measurement_unit
-      t.decimal :price
+      t.float :price
       t.integer :quantity
       
       t.timestamps
